Fixing a bug
Blender was crashing after remeshing landmark locations as vertices. The
bug so far seems to be coming from creating a bvhtree for the mesh that
has been added with new vertices, triangles, and edges. So removed the
usage of BVHTree. Yet to find out why BVH Tree causes Blender to crash