Code Improvement:
Every time the addConstraint method is called from the staticutilities
it calls the creation of visual markers
"bpy.ops.genericlandmarks.createconstraints". This is usually slower in
automated scenarios where the addConstraint needs to be called without
invoking the visual landmarks operator. So added a Flag
"create_visual_landmarks" as an argument to the method addConstraint
definition.