Bug fix and addition of a feature to SpectralOperators
*Fixed a bug of applyVertexColors and using histograms. The bug was
related to non-meeting of the if condition inside the method "getMinMax"
where if a mesh is having less vertices will not return the percentage
based min and max values using the histogram.

*Adding the histogram process as part of the existing operators of
GISIF, WKS, HKS with properties based on the mesh. The mesh now supports
the property of using histogram filterin colors, if true also has two
other properties that will be used as min and max percent of histogram
to use