Feature addition
Adding a new feature called mapping. This is useful in deformation
transfer, texture transfer, skeleton transfer, and other related
attributes transfer. Either a file can be given as input that contains
the mapping where each line is the index of the vertex for the mesh that
needs mapping, and each line has either 4 or 6 values. In the case of 4
values then the first value is an integer implyin the face id followed
by 3 float values implying the barycentric coordinate of the vertices in
that face by its looping order. If there are 6 values in each line then
the first 3 are the vertex ids and the following 3 are their respective
barycentric coordinates.