Updates
Adding Vertex to vertex mapping input between meshes. Each line in the
csv corresponds to a vertex on the other mesh. If mesh has 'n' vertices
then the csv file will have 'n' lines with values that are actually the
vertex indices of the othermesh it is mapped to. Indexing starts with
'0'. Ensure this is maintained if you are importing the mapping from
matlab